[sw-sysemu] Introduce ERBIUM simulator differentiation.
This patch provides the groundwork for ERBIUM simulator development.

At this moment, it provides a different memory map, disables the
service processor (and the IO shire in general) and sets a single
compute shire with eight minions.

This is not meant to be a complete nor precise emulation by now,
just the building block from which to build the actual emulator.

Notes: the ESRs are aligned to 2GB in Erbiu, opposed to the 4Gb
alignment in ETSOC1. Current memory map include that.
Also, we could be more C++ about this, but this allow not to have
expensive refactoring that affects current ETSOC1 emulation while
providing clear separation of devices and memory map.